fixed Tests. Introduce jacoco plugin

+++ src/main/java/de/crazymonkey/finanzinformation/repository/ShareRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import de.crazymonkey.finanzinformation.entity.Share;
 
-public interface ShareRepository extends CrudRepository<Share, Long> {
+public interface ShareRepository extends CrudRepository<Share, Integer> {
 
 	public List<Share> findBySymbol(String symbol);
 }
